+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file. This change
 - renamed some fn arguments for clarity
 - updated docstrings
 - encoding the default coordinates as 2d rather than nil
+- made sure the `plotly/plot` function would accept only one arity
 
 ## [-beta-1] - 2024-12-01
 - replaced the ad-hoc `WrappedValue` daratype with the Clojure `delay` idiom